C0227242300
- continue coding exercise 127 and updating the exercise template.
